The FastpData Object Class

The FastpData Object Class

FastpData(x)

Arguments

x

Path to a single zip archive or extracted folder for a individual fastp report.

Value

An object of class FastpData

Details

This object class is the main object required for generating plots and tables. Instantiation will first check for a .json file with the correct data structure, and will then parse all the data into R as a FastpData object. Fastp modules are contained as individual slots, which can be viewed using slotNames. Sub-modules are also contained within many larger modules with modules being based on the sections within a fastp html report

Individual modules can be returned using the function getModule() and specifying which module/sub-module is required. See getModule() for more details.

Slots

Summary

Contains three submodules 1) Before_filtering, 2) After_filtering and 3) Filtering_result. All values presented in the initial table for individual fastp reports are contained in other sections of the report

Adapters

Contains a tibble with all data from this module

Duplication

Contains a tibble with all duplication results

Insert_size

Contains a tibble with all insert size estimates

Before_filtering,After_filtering

The modules can be selected for either Read1 or Read2

paired

logical(1) indicating whether the file is from paired-end sequencing

command

character(1) with the executed command

version

character(1) with the fastp version being used

path

Path to the Fastp report