The FastqcData Object Class

The FastqcData Object Class

FastqcData(x)

Arguments

x

Path to a single zip archive or extracted folder for a individual FastQC report.

Value

An object of class FastqcData

Details

This object class is the main object required for generating plots and tables. Instantiation will first test for a compressed file (or extracted directory) with the correct data structure, and will then parse all the data into R as a FastqcData object. FastQC modules are contained as individual slots, which can be viewed using slotNames.

Individual modules can be returned using the function getModule() and specifying which module is required. See getModule() for more details.

Slots

Summary

Summary of PASS/WARN/FAIL status for each module

Basic_Statistics

The Basic_Statstics table from the top of a FastQC html report

Per_base_sequence_quality

The underlying data from the Per_base_sequence_quality module

Per_sequence_quality_scores

The underlying data from the Per_sequence_quality_scores module

Per_base_sequence_content

The underlying data from the Per_base_sequence_content module

Per_sequence_GC_content

The underlying data from the Per_sequence_GC_content module

Per_base_N_content

The underlying data from the Per_base_N_content module

Sequence_Length_Distribution

The underlying data from the Sequence_Length_Distribution module

Sequence_Duplication_Levels

The underlying data from the Sequence_Duplication_Levels module

Overrepresented_sequences

The underlying data from the Overrepresented_sequences module

Adapter_Content

The underlying data from the Adapter_Content module

Kmer_Content

The underlying data from the Kmer_Content module

Total_Deduplicated_Percentage

Estimate taken from the plot data for Sequence_Duplication_Levels. Only included in later versions of FastQC

version

The version of FastQC used for generation of the report (if available)

path

Path to the FastQC report

Examples

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)[1]

# Load the FASTQC data as a FastqcData object
fd <- FastqcData(fl)
fd
#> FastqcData for ATTG_R1.fastq 
#> Source Fastq file contains 24,978 reads.
#> Source FastQC file is located in /__w/_temp/Library/ngsReports/extdata/ATTG_R1_fastqc.zip