Perform the checks and return the labels

Checks for the presence of labels and returns defaults

.makeLabels(
  x,
  labels,
  pattern = ".(fast|fq|bam|sam|cram).*",
  col = "Filename",
  ...
)

Arguments

x

A data.frame with a column titled "Filename"

labels

Named vector of labels for plotting

pattern

character Regular expression to remove from filenames

col

character Column to use for generating labels

...

Not used

Value

Named character vector

Details

Takes a named vector of labels and checks for the correct fields. If no vector is supplied, returns the file names missing the specified pattern, which defaults to removing the suffixes fastq(.gz), fq(.gz), bam, sam or cram.

Examples

f <- paste0(c("File1", "File2"), ".fastq")
df <- data.frame(Filename = f, stringsAsFactors = FALSE)
ngsReports:::.makeLabels(df)
#> File1.fastq File2.fastq 
#>     "File1"     "File2"