Theoretical GC content

This object contains the theoretical GC content for each provided species, for both the genome and transcriptome, where available.

gcTheoretical

Format

An object of class TheoreticalGC of length 1.

Details

The object is defined with the S4 class TheoreticalGC. Species for which information is available can be found using the command gcAvail(gcTheoretical) and selecting the appropriate type.

Metadata is accessible using mData(gcTheoretical).

All GC content was calculated using code from https://github.com/mikelove/fastqcTheoreticalGC using BSgenome packages. This provides a default set of GC content data for common organisms generated using 100bp reads/fragments and 1e6 reads.

See also

Examples

## Check which genomes are included
gcAvail(gcTheoretical, "Genome")
#> # A tibble: 24 × 4
#>    Name          Group    Source Version   
#>    <chr>         <chr>    <chr>  <chr>     
#>  1 Alyrata       Plants   JGI    V1.0      
#>  2 Amellifera    Animals  UCSC   apiMel2   
#>  3 Athaliana     Plants   TAIR   Arapot11  
#>  4 Btaurus       Animals  UCSC   bosTau8   
#>  5 Celegans      Animals  UCSC   ce11      
#>  6 Cfamiliaris   Animals  UCSC   canFam3   
#>  7 Dmelanogaster Animals  UCSC   dm6       
#>  8 Drerio        Animals  UCSC   danRer10  
#>  9 Ecoli         Bacteria NCBI   2008/08/05
#> 10 Gaculeatus    Other    UCSC   gasAcu1   
#> # … with 14 more rows

## Check which transcriptomes are included
gcAvail(gcTheoretical, "Transcriptome")
#> # A tibble: 13 × 4
#>    Name          Group   Source Version 
#>    <chr>         <chr>   <chr>  <chr>   
#>  1 Athaliana     Plants  TAIR   Arapot11
#>  2 Btaurus       Animals UCSC   bosTau8 
#>  3 Celegans      Animals UCSC   ce11    
#>  4 Cfamiliaris   Animals UCSC   canFam3 
#>  5 Dmelanogaster Animals UCSC   dm6     
#>  6 Drerio        Animals UCSC   danRer10
#>  7 Hsapiens      Animals UCSC   hg38    
#>  8 Mmulatta      Animals UCSC   rheMac8 
#>  9 Mmusculus     Animals UCSC   mm10    
#> 10 Ptroglodytes  Animals UCSC   panTro5 
#> 11 Rnorvegicus   Animals UCSC   rn6     
#> 12 Scerevisiae   Yeast   UCSC   sacCer3 
#> 13 Sscrofa       Animals UCSC   susScr3