Get the maximum Adapter Content

Get the maximum Adapter Content across one or more FASTQC reports

maxAdapterContent(x, asPercent = TRUE)

Arguments

x

Can be a .FastqcFile, FastqcData, FastqcDataList or path

asPercent

logical. Format the values as percentages with the added \% symbol

Value

A tibble object containing the percent of reads with each adapter type at the final position

Details

This will extract the Adapter_Content module from the supplied object, and provide a tibble with the final value for each file.

Examples

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Get the maxAdapterContent
maxAdapterContent(fdl)
#> # A tibble: 6 × 4
#> # Groups:   Filename [6]
#>   Filename  Illumina_Small_RNA_A…¹ Illumina_Universal_A…² Nextera_Transposase_…³
#>   <chr>     <chr>                  <chr>                  <chr>                 
#> 1 ATTG_R1.… 0.00%                  47.23%                 0.00%                 
#> 2 ATTG_R2.… 0.00%                  56.24%                 0.00%                 
#> 3 CCGC_R1.… 0.00%                  15.76%                 0.00%                 
#> 4 CCGC_R2.… 0.00%                  20.49%                 0.00%                 
#> 5 GACC_R1.… 0.00%                  29.83%                 0.00%                 
#> 6 GACC_R2.… 0.00%                  37.91%                 0.00%                 
#> # ℹ abbreviated names: ¹​Illumina_Small_RNA_Adapter,
#> #   ²​Illumina_Universal_Adapter, ³​Nextera_Transposase_Sequence