Write an HTML Summary Report

Compiles an HTML report using a supplied template

writeHtmlReport(
  fastqcDir,
  template,
  outDir,
  usePlotly = TRUE,
  species = "Hsapiens",
  gcType = c("Genome", "Transcriptome"),
  nOver = 30,
  targetsDF,
  overwrite = FALSE,
  quiet = TRUE
)

Arguments

fastqcDir

A directory containing zipped, or extracted FastQC reports

template

The template file which will be copied into fastqcDir

outDir

The directory to write the compiled document to

usePlotly

Generate interactive plots?

species

Species/closely related species of sequenced samples

gcType

Is the data "Transcriptomic" or "Genomic" in nature?

nOver

The maximum number of Overrepresented Sequences to show

targetsDF

A data.frame with at least two columns named Filename and Label. The filenames should match the original fastq files, and the labels should be simply alternative labels for these files for convenience.

overwrite

logical. Overwrite any previous copies of the template file in the destination directory

quiet

logical. Show or hide markdown output in the Console.

Value

Silently returns TRUE and will output a compiled HTML file from the supplied Rmarkdown template file

Details

This will take a user supplied template, or the file supplied with the package and create an HTML summary of all standard FASTQC plots for all files in the supplied directory.

Examples

if (FALSE) {
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names= TRUE)
# Copy these files to tempdir() to avoid overwriting
# any files in the package directory
file.copy(fileList, tempdir(), overwrite = TRUE)
writeHtmlReport(fastqcDir = tempdir())
}